refactor: standardize naming conventions across services and entities
- rename EmployeeServices to EmployeeService for consistency
- standardize file names for SecurityConfig and NotFoundException
- relocate Role entity to a dedicated file for better organization
